tipsy-account/src/main/java/com/ssafy/tipsy/auth/PrincipalDetails.java
package com.ssafy.tipsy.auth;

import com.ssafy.tipsy.auth.model.UserAuth;
import lombok.Getter;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import java.util.ArrayList;
import java.util.Collection;

@Getter
public class PrincipalDetails implements UserDetails {

    private UserAuth userAuth;

    public PrincipalDetails(UserAuth userAuth) {
        this.userAuth = userAuth;
    }

    @Override
    public Collection<? extends GrantedAuthority> getAuthorities() {
        Collection<GrantedAuthority> collection = new ArrayList<>();
        collection.add(new GrantedAuthority() {
            @Override
            public String getAuthority() {
                return userAuth.getRole().toString();
            }
        });
        return collection;
    }

    @Override
    public String getPassword(){
        return "";
    }

    @Override
    public String getUsername() {
        return userAuth.getNickname();
    }

    @Override
    public boolean isAccountNonExpired() {
        return false;
    }

    @Override
    public boolean isAccountNonLocked() {
        return false;
    }

    @Override
    public boolean isCredentialsNonExpired() {
        return false;
    }

    @Override
    public boolean isEnabled() {
        return true;
    }
}